+++ python/chronos/dev/test/run-pytests-tf1.sh
@@ -30,8 +30,7 @@ ray stop -f
 echo "Running chronos tests TF1 and Deprecated API"
 python -m pytest -v test/bigdl/chronos/forecaster/test_mtnet_forecaster.py \
                     test/bigdl/chronos/model/test_mtnet.py \
-                    test/bigdl/chronos/model/test_Seq2Seq.py \
-                    test/bigdl/chronos/model/test_VanillaLSTM.py \
+                    test/bigdl/chronos/model/tf1 \
                     test/bigdl/chronos/autots/deprecated
 
 exit_status_0=$?
