+++ python/chronos/src/bigdl/chronos/autots/deprecated/model/time_sequence.py
@@ -16,8 +16,8 @@
 from copy import deepcopy
 
 from bigdl.orca.automl.model.abstract import BaseModel, ModelBuilder
-from bigdl.chronos.model.tf1_model._VanillaLSTM_keras import VanillaLSTM
-from bigdl.chronos.model.tf1_model._Seq2Seq_kears import LSTMSeq2Seq
+from bigdl.chronos.model.tf1.VanillaLSTM_keras import VanillaLSTM
+from bigdl.chronos.model.tf1.Seq2Seq_keras import LSTMSeq2Seq
 from bigdl.chronos.model.MTNet_keras import MTNetKeras
 from bigdl.chronos.autots.deprecated.feature.utils import save_config
 from bigdl.chronos.autots.deprecated.feature.time_sequence import TimeSequenceFeatureTransformer
