+++ python/chronos/test/bigdl/chronos/model/tf2/test_Seq2Seq_keras.py
@@ -20,7 +20,7 @@ import keras
 import pytest
 
 from bigdl.orca.test_zoo_utils import ZooTestCase
-from bigdl.chronos.model.Seq2Seq_keras import model_creator
+from bigdl.chronos.model.tf2.Seq2Seq_keras import LSTMSeq2Seq, model_creator
 import numpy as np
 
 
@@ -58,15 +58,20 @@ class TestSeq2Seq(ZooTestCase):
                        validation_data=self.test_data)
         yhat = self.model.predict(self.test_data[0])
         self.model.evaluate(self.test_data[0], self.test_data[1])
-        assert yhat.shape == self.test_data[1].shape
+        assert yhat.shape == (400, self.train_data[-1].shape[1], 2)
 
     def test_seq2seq_save_load(self):
         checkpoint_file = tempfile.TemporaryDirectory().name
+        self.model.fit(self.train_data[0],
+                       self.train_data[1],
+                       epochs=2,
+                       validation_data=self.test_data)
         self.model.save(checkpoint_file)
-        restore_model = keras.models.load_model(checkpoint_file)
+        restore_model = keras.models.load_model(checkpoint_file, custom_objects={"LSTMSeq2Seq": LSTMSeq2Seq})
         model_res = self.model.evaluate(self.test_data[0], self.test_data[1])
         restore_model_res = restore_model.evaluate(self.test_data[0], self.test_data[1])
         np.testing.assert_almost_equal(model_res, restore_model_res, decimal=5)
+        assert isinstance(restore_model, LSTMSeq2Seq)
     
 
 if __name__ == '__main__':
